requirements.txt
# Core ML/AI Libraries
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Deep Learning & Neural Networks
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Time Series Analysis
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0

# Geospatial & Satellite Data
geopandas>=0.13.0
rasterio>=1.3.0
folium>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
earthpy>=0.9.0
# satellite-image-deep-learning>=0.1.0

# Web Framework & API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0

# Data Processing
scipy>=1.11.0
xarray>=2023.6.0
netcdf4>=1.6.0
h5py>=3.9.0

# Web Scraping
beautifulsoup4>=4.12.0
scrapy>=2.10.0
selenium>=4.11.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0
configparser>=5.3.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
python-dateutil>=2.8.0
pytz>=2023.3

# Weather & Climate Data
meteostat>=1.6.0
wetterdienst>=0.60.0

# Visualization
bokeh>=3.2.0
dash>=2.12.0
streamlit>=1.25.0

# API Documentation
fastapi-users>=12.1.0
python-multipart>=0.0.6

# Async Support
aiofiles>=23.1.0
asyncio>=3.4.3

# Logging
loguru>=0.7.0

# Performance
numba>=0.57.0
dask>=2023.7.0
